requirements.txt
# PdfToTxt — 의회문서 PDF→OCR→LLM 요약 파이프라인
# 버전은 py310 환경 pip freeze 기준 고정

# === API 서버 ===
fastapi==0.131.0
uvicorn[standard]==0.41.0

# === PDF · 이미지 ===
pymupdf>=1.23.0
Pillow==12.1.1
numpy==2.2.6
opencv-python-headless==4.13.0.92

# === OCR (Tesseract) ===
pytesseract==0.3.13

# === 설정 · 환경 ===
pydantic-settings==2.13.1
python-dotenv==1.2.1

# === LLM 요약 (API 호출) ===
httpx==0.28.1
openai==2.21.0

# === 파일 업로드 ===
python-multipart==0.0.22
